[ADD] user and entity relationship

+++ app/views/order.py
@@ -2,7 +2,7 @@
 
 from flask import json, request
 from flask_restful import Resource
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 from app.models import db, Order, Menu, Meal, User
 from .validators import require_admin, num_check, name_validator
@@ -18,12 +18,14 @@ class Orders(Resource):
         """
         Method gets a order by id.
         """
+        current_user = User.query.filter_by(email=get_jwt_identity()).first()
         order = Order.query.filter_by(id=id).first()
         if order is None:
-            return {"status": "Failed!",
-                    "message": "Order  does not exist."}, 404
-        response = order.json_dump()
-        return response
+            return {"status": "Failed!", "message": "Order  does not exist."}, 404
+        if order.user_id == current_user.id:
+            response = order.json_dump()
+            return response
+        return {"message": "Order does not exist"}, 404
 
     @jwt_required
     def put(self, id):
@@ -31,31 +33,26 @@ class Orders(Resource):
         Method updates order.
         """
         json_data = request.get_json(force=True)
+        current_user = User.query.filter_by(email=get_jwt_identity()).first()
+        item = Menu.query.filter_by(id=json_data['item_id']).first()
         order = Order.query.filter_by(id=id).first()
         if order is None:
             return {"status": "Failed!",
                     "message": "Order does not exist."}, 404
-        if 'user_id' not in json_data or \
-                'item_id' not in json_data or 'quantity' not in json_data:
-            return {"status": "Failed!",
-                    "message": "Please supply user id, item id and quantity"}, 406
         if not num_check(json_data['item_id']) or not num_check(json_data['quantity']):
             return {"status": "Failed!", "meassage": "Meal_id,item_id and quantity must be integers."}, 406
-        item = Menu.query.filter_by(id=json_data['item_id']).first()
         if item is None:
-            return {"status": "Failed!",
-                    "message": "Item does not exist."}, 404
-        user = User.query.filter_by(id=json_data['user_id']).first()
-        if user is None:
-            return {"status": "Failed!!",
-                    "message": "User does not exist.Please enter a valid User id"}, 404
-        else:
-            order.user_id = json_data['user_id']
-            order.item_id = json_data['item_id']
-            order.quantity = json_data['quantity']
-            db.session.commit()
-            response = order.json_dump()
-            return{"status": "success", "data": response}, 200
+            return {"status": "Failed!", "message": "Item does not exist."}, 404
+
+        if item.user_id == current_user.id:
+            if order.user_id == current_user.id:
+                order.item_id = json_data['item_id']
+                order.quantity = json_data['quantity']
+                db.session.commit()
+                response = order.json_dump()
+                return{"status": "success", "data": response}, 200
+            return {"message": "Order does not exist"}, 404
+        return {"message": "Item does not exist!"}, 404
 
     @jwt_required
     def delete(self, id):
@@ -64,14 +61,16 @@ class Orders(Resource):
         """
         json_data = request.get_json(force=True)
         order = Order.query.filter_by(id=id).first()
+        current_user = User.query.filter_by(email=get_jwt_identity()).first()
         if order is None:
             return {"status": "Failed!!",
                     "data": "Order id does not exist.Please enter a valid order id"}, 404
-        else:
+        if order.user_id == current_user.id:
             Order.query.filter_by(id=id).delete()
             db.session.commit()
             response = json.loads(json.dumps(json_data))
             return {"status": "deleted!", "data": response}, 200
+        return {"message": "Order does not exist."}, 404
 
 
 class OrderList(Resource):
@@ -84,7 +83,10 @@ class OrderList(Resource):
         """
         Method to get all orders.
         """
-        orders = Order.query.all()
+        current_user = User.query.filter_by(email=get_jwt_identity()).first()
+        orders = Order.query.filter_by(user_id=current_user.id)
+        if orders is None:
+            return{"message": "You do not have orders at the moment."}, 404
         response = [order.json_dump() for order in orders]
         return {"status": "success", "data": response}, 200
 
@@ -94,34 +96,26 @@ class OrderList(Resource):
          Method creates an order.
         """
         json_data = request.get_json(force=True)
-        if 'user_id' not in json_data or \
-                'item_id' not in json_data or 'quantity' not in json_data:
+        current_user = User.query.filter_by(email=get_jwt_identity()).first()
+        if 'item_id' not in json_data or 'quantity' not in json_data:
             return {"status": "Failed!",
                     "message": "Please supply user id, item id and quantity"}, 406
         if not num_check(json_data['item_id']) or not num_check(json_data['quantity']):
-            return {"status": "Failed!"}
+            return {"status": "Failed!", "message": "Item and quantity id must be integers"}, 406
+        user_id = current_user.id
         item_id = json_data['item_id']
-        user_id = json_data['user_id']
         quantity = json_data['quantity']
         item = Menu.query.filter_by(id=item_id).first()
-        user = User.query.filter_by(id=user_id).first()
         if item_id == '':
             return {"status": "Failed!",
                     "message": "Item can not be empty."}, 406
-        elif user_id == '':
-            return {"status": "Failed!",
-                    "data": "User id can not be empty.Please enter a valid user id"}
         elif quantity == '':
             return {"status": "Failed!",
                     "message": "Quantity can not be empty."}, 406
         if item is None:
             return {"status": "Failed!",
                     "data": "Item id does not exist."}, 404
-        if user is None:
-            return {"status": "Failed!",
-                    "message": "User does not exist."}, 404
-        else:
-            order = Order(user_id=user_id, item_id=item_id, quantity=quantity)
-            order.save()
-            response = json.loads(json.dumps(order.json_dump()))
-            return {"status": "success", "data": response}, 201
+        order = Order(user_id=user_id, item_id=item_id, quantity=quantity)
+        order.save()
+        response = json.loads(json.dumps(order.json_dump()))
+        return {"status": "success", "data": response}, 201
